pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openx-mcp"
version = "1.1.0"
description = "OpenX — AI-powered GitHub automation MCP server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp>=1.2.0",
  "httpx>=0.27",
  "pydantic>=2.6",
  "PyGithub>=2.3",
  "python-dotenv>=1.0",
  "anthropic>=0.25",
  "langchain-core>=0.3",
  "langchain-anthropic>=0.3",
  "langgraph>=0.2",
  "langgraph-prebuilt>=0.1",
]

[project.scripts]
openx = "openx_agent.server:main"

[tool.setuptools.package-dir]
openx_agent = "openx-agent"
"openx_agent.analysis" = "openx-agent/analysis"
"openx_agent.tools" = "openx-agent/tools"

[tool.setuptools]
packages = ["openx_agent", "openx_agent.analysis", "openx_agent.tools"]
